componentize feature structure
