+++ src/components/NodeOperatorsTableCondensed.vue
@@ -1,0 +1,11 @@
+<template>
+  <table>a</table>
+</template>
+<script setup>
+defineProps({
+  nodeOperators: {
+    type: Array,
+    required: true,
+  },
+})
+</script>
